doc-engine/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
requests
google-auth
google-api-python-client